requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beanie==1.29.0
certifi==2025.4.26
click==8.1.8
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
lazy-model==0.2.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mirakuru==2.6.0
mongo-migrate==0.1.2
motor==3.7.0
packaging==25.0
pluggy==1.5.0
port-for==0.7.4
prometheus-client>=0.21.0
psutil==7.0.0
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
pymongo==4.12.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mongo==3.2.0
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
rich==14.0.0
rich-toolkit==0.14.5
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
toml==0.10.2
typer==0.15.3
typing-extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1

# OpenTelemetry for tracing and metrics
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp>=1.25.0
opentelemetry-exporter-otlp-proto-grpc>=1.25.0
opentelemetry-exporter-otlp-proto-http>=1.25.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-asgi>=0.45b0
